requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.1
pydantic-settings==2.2.1
PyYAML==6.0.1
redis==5.0.4
pytest==8.2.0
pytest-asyncio==0.23.6
httpx==0.27.0
asgi-lifespan==2.1.0